# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CyberArk Identity Sample NodeJS Server",
        "version": "1.0.0",
        "title": "NodeJS API",
        "contact": {
            "email": "abc@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http"],
    "host": "localhost:2200",
    "basePath": "/api",
    "paths" : {
...
        "/users/changePassword": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Changes the logged in user's password",
                "description": "Changes the logged in user's password",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "oldPassword": {
                                    "type": "string"
                                },
                                "newPassword": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/restResponse"
                        }
                    }
                }
            }
        },
        "/users/attributes/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user's attributes based on user's UUID or username",
                "description": "Get user's attributes based on user's UUID or username",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "username or user's UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/restResponse"
                        }
                    }
                }
            }
        },
...
        "/users/verifyTotp": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Verifies the TOTP provided by the user",
                "description": "Verifies the TOTP provided by the user",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "uuid": {
                                    "type": "string"
                                },
                                "otpCode": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/restResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "restResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "Result": {
                    "type": "object"
                },
                "Message": {
                    "type": "string"
                },
                "MessageID": {
                    "type": "string"
                },
                "Exception": {
                    "type": "string"
                },
                "ErrorID": {
                    "type": "string"
                },
                "ErrorCode": {
                    "type": "string"
                },
                "IsSoftError": {
                    "type": "string"
                },
                "InnerExceptions": {
                    "type": "string"
                }
            }
        },
...
    }
}